Update licenses and configurations, fix contract bug, and refactor tests

+++ contracts/MyERC721.sol
@@ -5,7 +5,7 @@ import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
 contract MyERC721 is ERC721, Ownable {
-  uint public tokenId;
+  uint256 public tokenId;
 
   constructor(
     string memory name,
